+++ src/main/java/command/history/UserHistoryCommand.java
@@ -1,4 +1,4 @@
-package command;
+package command.history;
 
 import com.jagrosh.jdautilities.command.Command;
 import com.jagrosh.jdautilities.command.CommandEvent;
